add SAGA verification
